Chore: refactor goal to goal-list
