+++ src/app/app.component.html
@@ -1,5 +1,5 @@
 <h1>
   {{title}}
 </h1>
-<goal> </goal>
+<goal-list></goal-list>
 <router-outlet></router-outlet>
