+++ src/app/app.module.ts
@@ -6,7 +6,7 @@ import { HttpModule } from '@angular/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { GoalComponent } from './goal/goal.component';
+import { GoalComponent } from './goal-list/goal-list.component';
 
 @NgModule({
   declarations: [
